requirements.txt
asgiref==3.7.2
astroid==3.1.0
colorama==0.4.6
dill==0.3.8
Django==5.0.3
django-filter==23.5
djangorestframework==3.14.0
Faker==24.0.0
isort==5.13.2
lazy-object-proxy==1.10.0
mccabe==0.7.0
platformdirs==4.2.0
pylint==3.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
tomlkit==0.12.4
typed_ast==1.5.5
tzdata==2024.1
validate-docbr==1.10.0
wrapt==1.16.0
